Created common abstract class AbstractPoint for Point and Center which also fixed some tests' breakage.
